% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diallel_sim.R
\name{simulate.prior.diallel}
\alias{simulate.prior.diallel}
\title{Simulate diallel data from the diffuse priors of the BayesDiallel model}
\usage{
\method{simulate}{prior.diallel}(M = 8, n.each = 5, num.sim = 1,
  strains = LETTERS[1:M], hyper.ga.alpha = 0.002, hyper.ga.beta = 2)
}
\arguments{
\item{M}{DEFAULT: 8. The number of strains in the diallel cross. Default of 8 is the number of founders in the Collaborative Cross.}
}
\description{
This function generates data from the BayesDiallel model, based on the default diffuse priors. 
It does not currently allow for sex interactions with strain-level effects, as is possible 
in the proper BayesDiallel package.
}
\examples{
simulate.prior.diallel()
}
